requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pdfminer.six==20200517
pycryptodome==3.9.8
requests==2.24.0
sortedcontainers==2.2.2
urllib3==1.25.9
